% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calfire_graph.R
\name{calfire_graph}
\alias{calfire_graph}
\title{California Wildfire Graph}
\usage{
calfire_graph(fire_data, metric)
}
\arguments{
\item{fire_data}{The input dataset for calculating summary statistics (Data is provided in the 'climate' package and is titled 'calfires')}

\item{metric}{The metric statistic that the graph will show (must be an integer between 1 - 6)
\itemize{
\item 1: Total fires per year
\item 2: Total acres burned per year
\item 3: Total injuries sustained due to fires per year
\item 4: Total fatalies due to fires per year
\item 5: Total structures damaged due to fires per year
\item 6: Total structures destroyed due to fires per year
}}
}
\value{
Bar graph of chosen fire statistic per year from 2013 - 2018
}
\description{
This function produces a graph of summary statistics, by year, for California Wildfires from 2013 - 2019
}
\examples{
calfire_graph(fire_data = calfires, metric = 1)
}
\author{
Chase Brewster
}
