% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calfire_stats.R
\name{calfire_stats}
\alias{calfire_stats}
\title{California Wildfire Stats}
\usage{
calfire_stats(fire_data, year)
}
\arguments{
\item{fire_data}{The input dataset for calculating summary statistics (Data is provided in the 'climate' package and is titled 'calfires')}

\item{year}{The year for which the function will return statistics (must be between 2013 - 2019)}
}
\value{
List of the number of fires, total acres burned, injuries, fatalities, structures damaged, and structures destroyed, for a given year
}
\description{
This function produces summary statistics, by year, for California Wildfires from 2013 - 2019
}
\examples{
calfire_stats(fire_data = calfires, year = 2018)
}
\author{
Chase Brewster
}
