% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calfires.R
\docType{data}
\name{calfires}
\alias{calfires}
\title{California Wildfire Incidents}
\format{A data frame with 1636 rows (fires) and 40 columns (variables).
Example variables:
\describe{
 \item{ArchiveYear}{the year the fire occured}
 \item{AcresBurned}{total acreage burned during the fire}
 \item{Extinguished}{date and time the fire was considered extinguished}
 \item{Name}{the name of the fire}
 \item{Helicopters}{number of helicopters used to extinguish the fire}
 \item{Injuries}{number of injuries sustained due to the fire}
}}
\source{
: Kaggle \url{https://www.kaggle.com/ananthu017/california-wildfire-incidents-20132020/data}
}
\usage{
calfires
}
\description{
A record of fire incidents in California from 2013 - 2020.
The dataset contains over 1600 wildfires.
}
\keyword{datasets}
