% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wl.R
\name{chartWL}
\alias{chartWL}
\title{Plots a climate chart}
\usage{
chartWL(adata)
}
\arguments{
\item{adata}{a data frame of daily weather data. First columns include locality data}
}
\value{
nothing
}
\description{
Plots Walter-Lieth climate chart. Uses a base function from climatol package and
adds some more details. Work in progress.
}

