% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_hydro.R
\name{clean_metadata_hydro}
\alias{clean_metadata_hydro}
\title{Hydrological metadata cleaning (IMGW-PIB data only)}
\usage{
clean_metadata_hydro(address, interval)
}
\arguments{
\item{address}{URL address of the metadata file}

\item{interval}{temporal interval}
}
\description{
Internal function for hydrological metadata cleaning
}
\keyword{internal}
