% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-package.R
\docType{package}
\name{climate-package}
\alias{climate}
\alias{climate-package}
\title{climate: Meteorological and Hydrological Data}
\description{
Download meteorological and hydrological data from the Institute of Meteorology and Water Management - National Research Institute (<https://dane.imgw.pl/>).
    EXPANDEXPANDEXPAND.
    This package also allows for adding geographical coordinates for each observation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bczernecki/climate}
  \item Report bugs at \url{https://github.com/bczernecki/climate/issues}
}

}
\author{
\strong{Maintainer}: Bartosz Czernecki \email{nwp@amu.edu.pl} (0000-0001-6496-1386)

Authors:
\itemize{
  \item Arkadiusz Głogowski \email{arkadiusz.glogowski@upwr.edu.pl} (0000-0002-7587-8892)
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)
}

Other contributors:
\itemize{
  \item IMGW-PIB (source of the data) [contributor]
}

}
\keyword{internal}
