% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_demo.R
\docType{data}
\name{co2_demo}
\alias{co2_demo}
\title{Examplary CO2 dataset from Mauna Loa Observatory (NOAA dataset)}
\format{
An object of class \code{data.frame} with 745 rows and 7 columns.
}
\usage{
co2_demo
}
\description{
The object contains pre-downloaded CO2 dataset from Mauna Loa observatory
The snapshot was taken 2020/05/05.
}
\examples{
data(co2_demo)
head(co2_demo)
}
\keyword{datasets}
\keyword{meteo}
