% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gently.R
\name{download_gently}
\alias{download_gently}
\title{Download gently}
\usage{
download_gently(url, output)
}
\arguments{
\item{url}{character vector with URL to check}

\item{output}{character vector with URL to check}
}
\description{
Internal function for testing url/internet connection and stopping if url is wrong or no internet access detected
}
\examples{
\donttest{
 link = "https://www1.ncdc.noaa.gov/pub/data/noaa/2019/123300-99999-2019.gz"
 output = basename(link)
 climate::download_gently(url = link, output = output)
}

}
