% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_data.R
\docType{data}
\name{extent_data}
\alias{extent_data}
\title{Test data for sea ice extent}
\format{a data frame of 494 rows and 6 columns}
\source{
ftp://sidads.colorado.edu/DATASETS/NOAA/G02135/north/monthly/data/

\describe{
\item{year}{Year in four digit format}
\item{mo}{Month in numeric format 1-12}
\item{data-type}{Goddard: Sea Ice Concentrations from Nimbus-7 SMMR and DMSP SSM/I-SSMIS Passive Microwave Data}
\item{region}{Hemisphere N = Northern}
\item{extent}{Sea ice extent in millions of square kilometers; no data values -9999}
\item{area}{Sea ice area in millions of square kilometers; no data values -9999}
}
}
\usage{
extent_data
}
\description{
Sea ice extent (in square kilometers) each month from 1979-2019
}
\keyword{datasets}
