% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_all_station_names}
\alias{find_all_station_names}
\title{Find all variants of station' names}
\usage{
find_all_station_names(station_name)
}
\arguments{
\item{station_name}{character vector of station names}
}
\value{
character vector of station names with all variants of station's names
}
\description{
For IMGW-PIB stations different naming were used historically. For example,
\code{POZNAŃ} and ``POZNAŃ-ŁAWICA, thus both names should be used when searching
for the station. This function finds all variants of station' names
status information and expand the created object
}
\examples{
{
find_all_station_names(c("WARSZAWA", "POZNAŃ"))
}
}
