% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedamage.R
\name{firedamage}
\alias{firedamage}
\title{Wildfire Damage Calculator}
\usage{
firedamage(
  acres,
  value,
  structures,
  avgstructureval,
  helicopters,
  days,
  firecrews,
  injuries,
  fatalities
)
}
\arguments{
\item{acres}{The number of acres burned in the fire}

\item{value}{The average land value of the area burned (in USD per acre)}

\item{structures}{The number of structures severly damaged and destroyed in the fire}

\item{avgstructureval}{The average value of structures in the burn area (in USD per structure)}

\item{helicopters}{The number of helicopters used to fight the fire}

\item{days}{The number of days the fire burned}

\item{firecrews}{The number of firecrews that fought the fire}

\item{injuries}{The number of injuries sustained due to the fire}

\item{fatalities}{The number of fatalities due to the fire}
}
\value{
Total monetary damage of a wildfire, in USD
}
\description{
This function roughly calculates the monetary damage of a wildfire
}
\examples{
firedamage(acres = 100, value = 5000, structures = 20, avgstructureval = 100000, helicopters = 5, days = 5, firecrews = 8, injuries = 8, fatalities= 0)
}
\author{
Chase Brewster
}
