% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_abbrev.R
\docType{data}
\name{hydro_abbrev}
\alias{hydro_abbrev}
\title{Definitions of hydrological parameters used for shortening column names}
\format{The data contains a data.frame with ca. 20 elements described in three ways:
\itemize{
\item{fullname} {original column names as downloaded from the repository}
\item{abbr_eng} {shorten column names with abbreviations derived from the most popular scheme used for meteorological parameters}
\item{fullname_eng} {detailed description of downloaded meteorological variables}
}
The object is created mostly to be used altogether with the hydro_shortening() function}
\usage{
hydro_abbrev
}
\description{
The object contains 3 columns that are currently used for improving readability of the downloaded dataset:
fullname, abbr_eng, and fullname_eng
}
\examples{
data(hydro_abbrev)
head(hydro_abbrev)
}
\keyword{abbreviations}
\keyword{datasets}
\keyword{hydro}
\keyword{shortening}
