% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_imgw.R
\name{hydro_imgw}
\alias{hydro_imgw}
\title{Hydrological data from IMGW}
\usage{
hydro_imgw(interval, year, value = "H", station = NULL, ...)
}
\arguments{
\item{interval}{temporal resolution of the data ("daily" or "monthly")}

\item{year}{vector of years (e.g., 1966:2000)}

\item{value}{type of data (can be: state - "H" (default), flow - "Q", or
temperature - "T")}

\item{station}{vector of hydrological stations danepubliczne.imgw.pl;
can be given as station name with CAPITAL LETTERS (character)
It accepts either names (characters in CAPITAL LETTERS) or stations' IDs (numeric)}

\item{...}{other parameters that may be passed e.g. to the 'shortening'
function that shortens column names}
}
\value{
A data.frame with columns describing the hydrological parameters
(e.g. flow, water level) where each row represent a measurement,
depending on the interval, at a given hour, month or year.
IMGW parameter columns carry a \code{label} attribute with the original metadata label.
If \code{coords = TRUE} additional two columns with geographic coordinates are added.
}
\description{
Downloading daily, and monthly hydrological data from the measurement stations
available in the danepubliczne.imgw.pl collection
}
\examples{
\donttest{
  x = hydro_imgw("monthly", year = 1999)
  head(x)
}
}
