% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_metadata.R
\name{hydro_metadata}
\alias{hydro_metadata}
\title{Hydrological metadata}
\usage{
hydro_metadata(interval)
}
\arguments{
\item{interval}{temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")}
}
\description{
Downloading the description (metadata) to hydrological data available in the danepubliczne.imgw.pl repository.
By default, the function returns a list or data frame for a selected subset
}
\examples{
\donttest{
  meta <- hydro_metadata(interval = "daily")
  meta <- hydro_metadata(interval = "monthly")
  meta <- hydro_metadata(interval = "semiannual_and_annual")
}
}
