% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_metadata_imgw.R
\name{hydro_metadata_imgw}
\alias{hydro_metadata_imgw}
\title{Hydrological metadata}
\usage{
hydro_metadata_imgw(interval, allow_failure = TRUE)
}
\arguments{
\item{interval}{temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")}

\item{allow_failure}{logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE}
}
\description{
Downloading the description (metadata) to hydrological data available in the danepubliczne.imgw.pl repository.
By default, the function returns a list or data frame for a selected subset
}
\keyword{internal}
