% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_shortening_imgw.R
\name{hydro_shortening_imgw}
\alias{hydro_shortening_imgw}
\title{Shortening column names for hydrological variables}
\usage{
hydro_shortening_imgw(data, remove_duplicates = TRUE)
}
\arguments{
\item{data}{downloaded dataset with original column names}

\item{remove_duplicates}{whether to remove duplicated column names
(default TRUE - i.e., columns with duplicated names are deleted)}
}
\value{
data.frame with short English names of hydrological parameters.
Existing column attributes, including the original IMGW \code{label} metadata,
are preserved.
}
\description{
Shortening column names of hydrological parameters to improve the readability of downloaded dataset from
the danepubliczne.imgw.pl collection and removing duplicated column names
}
\examples{
\donttest{
  monthly = data = hydro_imgw("monthly", year = 1969)
  
  if (is.data.frame(monthly)) {
  abbr = hydro_shortening_imgw(data = monthly, remove_duplicates = TRUE)
  head(abbr)
  }
}

}
