% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgw_synop_stations.R
\docType{data}
\name{imgw_synop_stations}
\alias{imgw_synop_stations}
\title{Table with codes for SYNOP stations}
\format{The data contains a data.frame with 66 obs. of 2 variables:
\itemize{
\item{id} {Station ID}
\item{station} {Station name}
}}
\usage{
imgw_synop_stations
}
\description{
The object contains: Polish ID numbers (as used by Polish met service), and names of main (i.e. synop) station in Poland. The table is used for selecting choosen stations to be downloaded and thus saving internal memory.
}
\examples{
data(imgw_synop_stations)
head(imgw_synop_stations)
}
\keyword{datasets}
\keyword{meteo}
