% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march_extent.R
\docType{data}
\name{march_extent}
\alias{march_extent}
\title{Monthly sea ice extent test data}
\format{a data frame of 41 rows and 6 columns}
\source{
ftp://sidads.colorado.edu/DATASETS/NOAA/G02135/north/monthly/data/

\describe{
\item{year}{Year in four digit format}
\item{mo}{Month in numeric format 1-12}
\item{data-type}{Goddard: Sea Ice Concentrations from Nimbus-7 SMMR and DMSP SSM/I-SSMIS Passive Microwave Data}
\item{region}{Hemisphere N = Northern}
\item{extent}{Sea ice extent in millions of square kilometers}
\item{area}{Sea ice area in millions of square kilometers}
}
}
\usage{
march_extent
}
\description{
Sea ice extent for March (in square kilometers) from 1979-2019
}
\keyword{datasets}
