% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_abbrev.R
\docType{data}
\name{meteo_abbrev}
\alias{meteo_abbrev}
\title{Definitions of meteorological parameters used for shortening column names}
\format{The data contains a data.frame with ca. 250 elements described in three ways:
\itemize{
\item{fullname} {original column names as downloaded from the repository}
\item{abbr_eng} {shorten column names with abbreviations derived from the most popular scheme used for meteorological parameters}
\item{fullname_eng} {detailed description of downloaded meteorological variables}
}
The object is created mostly to be used altogether with the meteo_shortening function}
\usage{
meteo_abbrev
}
\description{
The object contains 3 columns that are currently used for improving readability of the downloaded dataset:
fullname, abbr_eng, and fullname_eng
}
\examples{
data(meteo_abbrev)
head(meteo_abbrev)
}
\keyword{abbreviations}
\keyword{datasets}
\keyword{meteo}
\keyword{shortening}
