% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_imgw_datastore.R
\name{meteo_imgw_datastore}
\alias{meteo_imgw_datastore}
\title{IMGW meteorological data from IMGW datastore repository}
\usage{
meteo_imgw_datastore(year, allow_failure = TRUE, ...)
}
\arguments{
\item{year}{vector of years (e.g., 1966:2000)}

\item{allow_failure}{logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE}

\item{...}{other parameters that may be passed to the 'shortening'
function that shortens column names}

\item{rank}{rank of the stations: "synop" (default), "climate", or "precip"}
}
\description{
Downloading hourly (meteorological) data from the telemetric stations
available in the danepubliczne.imgw.pl/datastore collection since 2008
}
\examples{
\donttest{
  #TODO
}

}
