% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_imgw_hourly.R
\name{meteo_imgw_hourly}
\alias{meteo_imgw_hourly}
\title{Hourly IMGW meteorological data}
\usage{
meteo_imgw_hourly(
  rank = "synop",
  year,
  status = FALSE,
  coords = FALSE,
  station = NULL,
  allow_failure = TRUE,
  ...
)
}
\arguments{
\item{rank}{rank of the stations: "synop" (default), "climate", or "precip"}

\item{year}{vector of years (e.g., 1966:2000)}

\item{status}{leave the columns with measurement and observation statuses
(default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}

\item{station}{name of meteorological station(s) (character vector)}

\item{allow_failure}{logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE}

\item{...}{other parameters that may be passed to the 'shortening'
function that shortens column names}
}
\value{
meteorological data for the hourly time interval
}
\description{
Downloading hourly (meteorological) data from the SYNOP / CLIMATE / PRECIP stations
available in the danepubliczne.imgw.pl collection
}
\examples{
\donttest{
hourly = meteo_imgw_hourly(rank = "climate", year = 1984)
head(hourly)
}

}
