% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_metadata_imgw.R
\name{meteo_metadata_imgw}
\alias{meteo_metadata_imgw}
\title{Meteorological metadata}
\usage{
meteo_metadata_imgw(interval, rank)
}
\arguments{
\item{interval}{temporal resolution of the data ("hourly", "daily", "monthly")}

\item{rank}{rank of station ("synop", "climate", "precip")}
}
\description{
Downloading the description (metadata) to the meteorological data available in the dane.imgw repository.imgw.pl.
By default, the function returns a list or data frame for a selected subset
}
\examples{
\donttest{
  meta <- climate:::meteo_metadata_imgw(interval = "hourly", rank = "synop")
  meta <- climate:::meteo_metadata_imgw(interval = "daily", rank = "synop")
  meta <- climate:::meteo_metadata_imgw(interval = "monthly", rank = "precip")
}
}
\keyword{internal}
