% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_shortening_imgw.R
\name{meteo_shortening_imgw}
\alias{meteo_shortening_imgw}
\title{Shortening column names for meteorological variables}
\usage{
meteo_shortening_imgw(data, remove_duplicates = TRUE)
}
\arguments{
\item{data}{downloaded dataset with original column names}

\item{remove_duplicates}{whether to remove duplicated column names (default TRUE - i.e., columns with duplicated names are deleted)}
}
\value{
data.frame with short English names of meteorological parameters.
Existing column attributes, including the original IMGW \code{label} metadata,
are preserved.
}
\description{
Shortening column names of meteorological parameters to improve the readability of downloaded dataset from the danepubliczne.imgw.pl collection and removing duplicated column names
}
\examples{
\donttest{
  monthly = meteo_imgw("monthly", rank = "climate", year = 1969)
  
  abbr = meteo_shortening_imgw(data = monthly, remove_duplicates = TRUE)
  head(abbr)
}

}
