% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_stations.R
\docType{data}
\name{meteo_stations}
\alias{meteo_stations}
\title{Location of the meteorological stations}
\format{The data contains a data.frame with 1998 obs. of 3 variables:
\itemize{
\item{id} {Station ID}
\item{X} {Longitude}
\item{Y} {Latitude}
}
The object is in the geographic coordinates using WGS84 (EPSG:4326).}
\usage{
meteo_stations
}
\description{
The object contains weather stations
coordinates, ID numbers, and elevations
}
\examples{
data(meteo_stations)
head(meteo_stations)
}
\keyword{datasets}
\keyword{meteo}
