% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_stations_ogimet.R
\name{nearest_stations_ogimet}
\alias{nearest_stations_ogimet}
\title{Scrapping a list of meteorological (Synop) stations for a defined country from the Ogimet webpage}
\usage{
nearest_stations_ogimet(country = "United+Kingdom", date = Sys.Date(),
  add_map = FALSE, nearest = FALSE, point = c(0, 0),
  numbers_station = 1)
}
\arguments{
\item{country}{country name; for more than two words they need to be seperated with a plus character (e.g. "United+Kingdom")}

\item{date}{a day when measurements were done in all available locations}

\item{add_map}{logical - whether to draw a map with downloaded metadata (requires maps/mapdata packages)}

\item{point}{a vector of two coordinates (longitude, latitude) for a point we want to find nearest stations to (e.g. c(0, 0))}

\item{numbers_station}{numbers of the nearest stations to select}
}
\description{
Returns a list of meteorological stations with their coordinates from the ogimet webpage.
The returned list is valid only for a given day.
It is possible to select stations located closest to the indicated point.
}
\examples{
\donttest{
  stations_ogimet(country = "Australia", add_map = TRUE)
}

}
