% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogimet_stations.R
\name{ogimet_stations}
\alias{ogimet_stations}
\title{Scrapping list of meteorological (Synop) stations for a defined country from the Ogimet webpage}
\usage{
ogimet_stations(country = "United+Kingdom", date = Sys.Date(),
  add_map = FALSE)
}
\arguments{
\item{country}{country name; for more than 2 words seperated with plus character (e.g. "United+Kingdom")
numeric fields (logical value TRUE (default) or FALSE)}

\item{date}{a day when measurements were done in all available locations}

\item{add_map}{logical - whether to draw a map with downloaded metadata (requires maps/mapdata packages)}
}
\description{
Returns list of meteorological stations with its coordinates from the ogimet webpage. The returned list is valid only for a given day
}
\examples{
\donttest{
  ogimet_stations(country = "Australia", add_map = T)
}

}
