% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_demo.R
\docType{data}
\name{profile_demo}
\alias{profile_demo}
\title{Exemplary sounding profile from University of Wyoming dataset}
\format{
The data contains list of two data.frames as derived using sounding_wyoming() function
}
\usage{
profile_demo
}
\description{
The object contains pre-downloaded atmospheric (sounding) profile for Leba, PL rawinsonde station.
The measurement was taken 2000/03/23 at 00 UTC.
}
\examples{
data(profile_demo)
head(profile_demo)
}
\keyword{datasets}
\keyword{meteo}
