% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_extremes.R
\name{si_extremes}
\alias{si_extremes}
\title{Describe sea ice extent extremes in the Northern Hemisphere}
\usage{
si_extremes(si, showtable = FALSE)
}
\arguments{
\item{si}{data frame of sea ice data containing at least three columns: year, mo, extent}

\item{showtable}{displays optional table of outputs, default is false}
}
\value{
list with the following items
\describe{
\item{max}{Maximum sea ice extent in millions of squre kilometers}
\item{max_yr}{Year corresponding to the maximum sea ice extent}
\item{max_mo}{Month corresponding to the maximum sea ice extent}
\item{min}{Minimum sea ice extent in millions of square kilometers}
\item{min_yr}{Year corresponding to the minimum sea ice extent}
\item{min_mo}{Month corresponding to the minimum sea ice extent}
\item{si_table}{Table displaying the above output values, null if showtable=F}
}
}
\description{
Find sea ice extent maximum and minimum with corresponding year and month
}
\examples{
Generate some input data:
input <- data.frame(year = rep(1990, 12), mo = seq(1,12,1), extent = runif(12,2,15))
si_extremes(si = input, showtable = T)


}
\references{
Example data downloaded from [ftp://sidads.colorado.edu/DATASETS/NOAA/G02135/north/monthly/data/]
Metadata for package data: [https://nsidc.org/data/G02135/versions/3]
}
