% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sle.R
\name{sle}
\alias{sle}
\title{Describes the equivalent change in sea level rise from ice volume}
\usage{
sle(ice_vol, above)
}
\arguments{
\item{ice_vol}{a single numeric value sea ice volume in kilometers cubed}

\item{above}{the percent of the sea ice that is above the surface of the water, in decimal format}
}
\value{
the sea level rise equivalent (in mm) caused by the melting of the input volume of sea ice (ice_vol)
}
\description{
Calculate the sea level rise equivalent (in mm) from melting any given volume of sea ice
}
\examples{
sle(ice_vol = 500, above = 0.6)
}
