% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations_hydro_imgw_telemetry.R
\name{stations_hydro_imgw_telemetry}
\alias{stations_hydro_imgw_telemetry}
\title{IMGW hydrological telemetry stations}
\usage{
stations_hydro_imgw_telemetry()
}
\value{
data table with metadata for over 850 stations.
Metadata contains: station ID, station name, river, latitude, longitude, altitude
}
\description{
Retrieving current metadata for hydrological stations used in the telemetric systems of the IMGW-PIB datastore (danepubliczne.imgw.pl/datastore)
}
\examples{
\donttest{
  hydro_telemetry_stations = stations_hydro_imgw_telemetry()
}

}
