% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_imgw_telemetry_stations.R
\name{stations_meteo_imgw_telemetry}
\alias{stations_meteo_imgw_telemetry}
\title{IMGW telemetry stations}
\usage{
stations_meteo_imgw_telemetry()
}
\value{
data table with metadata for over 500 stations. Metadata contains: station ID, station name, river, latitude, longitude, altitude
}
\description{
Retrieving current metadata for stations used in the telemetric systems of the IMGW-PIB datastore (danepubliczne.imgw.pl/datastore)
}
\examples{
\donttest{
  telemetry_stations = meteo_imgw_telemetry_stations()
}

}
