% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_meteo.R
\name{clean_metadata_meteo}
\alias{clean_metadata_meteo}
\title{Meteorological metadata cleaning}
\usage{
clean_metadata_meteo(address, rank = "synop", interval = "hourly")
}
\arguments{
\item{address}{URL address of the metadata file}

\item{rank}{stations' rank}

\item{interval}{temporal interval}
}
\description{
Internal function for meteorological metadata cleaning
}
\examples{
\donttest{
  my_add = paste0("https://dane.imgw.pl/data/dane_pomiarowo_obserwacyjne/",
                  "dane_meteorologiczne/dobowe/synop/s_d_format.txt")
  climate:::clean_metadata_meteo(address = my_add, rank = "synop", interval = "hourly")
}

}
\keyword{internal}
