% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-package.R
\docType{package}
\name{climate-package}
\alias{climate}
\alias{climate-package}
\title{climate: Interface to Download Meteorological (and Hydrological) Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automatize downloading of meteorological and hydrological data from publicly available repositories: OGIMET (\url{http://ogimet.com/index.phtml.en}), University of Wyoming - atmospheric vertical profiling data (\url{http://weather.uwyo.edu/upperair/}), Polish Institute of Meteorology and Water Management - National Research Institute (\url{https://danepubliczne.imgw.pl}), and National Oceanic & Atmospheric Administration (NOAA). This package also allows for searching geographical coordinates for each observation and calculate distances to the nearest stations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bczernecki/climate}
  \item \url{https://bczernecki.github.io/climate/}
  \item Report bugs at \url{https://github.com/bczernecki/climate/issues}
}

}
\author{
\strong{Maintainer}: Bartosz Czernecki \email{nwp@amu.edu.pl} (\href{https://orcid.org/0000-0001-6496-1386}{ORCID})

Authors:
\itemize{
  \item Arkadiusz Głogowski \email{arkadiusz.glogowski@upwr.edu.pl} (\href{https://orcid.org/0000-0002-7587-8892}{ORCID})
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

Other contributors:
\itemize{
  \item IMGW-PIB (source of the data) [contributor]
}

}
\keyword{internal}
