\name{Climate-package}
\alias{Climate-package}
\alias{Climate}
\docType{package}
\title{
What the package does (short line)
	Climate data manipulation
}
\description{
More about what it does (maybe more than one line)
	Climate data manipulation. Particularly in the context of ecological modeling, such as crop growth simulation modeling.
}
\details{
\tabular{ll}{
Package: \tab Climate\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-10-17\cr
License: \tab GPL v3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: <r.hijmans@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\keyword{ spatial }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
