% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_metadata_imgw.R
\name{hydro_metadata_imgw}
\alias{hydro_metadata_imgw}
\title{Hydrological metadata}
\usage{
hydro_metadata_imgw(interval)
}
\arguments{
\item{interval}{temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")}
}
\description{
Downloading the description (metadata) to hydrological data available in the dane.imgw.pl repository.
By default, the function returns a list or data frame for a selected subset
}
\examples{
\donttest{
  meta <- climate:::hydro_metadata_imgw(interval = "daily")
  meta <- climate:::hydro_metadata_imgw(interval = "monthly")
  meta <- climate:::hydro_metadata_imgw(interval = "semiannual_and_annual")
}
}
\keyword{internal}
