% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_stations_ogimet.R
\name{nearest_stations_ogimet}
\alias{nearest_stations_ogimet}
\title{List of nearby synop stations for a defined geographical location}
\usage{
nearest_stations_ogimet(country = "United+Kingdom", date = Sys.Date(),
  add_map = FALSE, nearest = TRUE, point = c(50, 0),
  numbers_station = 1)
}
\arguments{
\item{country}{country name; for more than two words they need to be seperated with a plus character (e.g. "United+Kingdom")}

\item{date}{optionally, a day when measurements were done in all available locations; current Sys.Date used by default}

\item{add_map}{logical - whether to draw a map for a returned data frame (requires maps/mapdata packages)}

\item{nearest}{logical vector indicating whether to look only for the nearest stations or to all available stations in a country}

\item{point}{a vector of two coordinates (longitude, latitude) for a point we want to find nearest stations to (e.g. c(0, 0))}

\item{numbers_station}{how many nearest stations will be returned from the given geographical coordinates}
}
\description{
Returns a data frame of meteorological stations with their coordinates and distance from a given location based on the ogimet webpage.
The returned list is valid only for a given day.
}
\examples{
\donttest{
  nearest_stations_ogimet(country = "United+Kingdom", point = c(10,50), add_map = T, numbers_station = 60)
}

}
